test : 리팩토링으로 새롭게 분리된 메소드들에 대해 테스트 작성
Inputchecker.isValid()
RandomGenerator.generate()
stringUtil.isAlphabetString()
