+++ src/test/java/carRace/validations/InputCheckerTest.java
@@ -1,0 +1,14 @@
+package carRace.validations;
+
+import org.junit.jupiter.api.Test;
+
+public class InputCheckerTest {
+
+    @Test
+    public void isValidTest(){
+        String[] testStrings = new String[]{
+                "hello", "wolrd", "How", "Are", "you"
+        };
+        InputChecker.isValid(testStrings);
+    }
+}
